chore: relocate the paint palette
