+++ .changeset/fifty-moons-grab.md
@@ -1,0 +1,6 @@
+---
+'@seedcord/errors': patch
+'@seedcord/logger': patch
+---
+
+Moved `paint` to the errors package
