+++ packages/errors/src/internal.index.ts
@@ -1,2 +1,5 @@
 export { SeedcordError, SeedcordTypeError, SeedcordRangeError } from './SeedcordError';
+
+// these files don't exactly belong in this package but are used in places logger isn't imported
+export { paint } from './palette';
 export { validateDiscordToken } from './validateDiscordToken';
