+++ packages/logger/src/index.ts
@@ -2,7 +2,8 @@ export { FRAMEWORK_CHANNELS } from './channels';
 export { Logger } from './Logger';
 export { LoggerChannelRegistry } from './LoggerChannelRegistry';
 export { ObjectConsoleSink } from './ObjectConsoleSink';
-export { LEVEL_COLOR, paint } from './palette';
+export { paint } from '@seedcord/errors/internal';
+export { LEVEL_COLOR } from './palette';
 
 export type {
     FrameworkChannel,
