+++ packages/logger/src/palette.ts
@@ -1,5 +1,3 @@
-import chalk from 'chalk';
-
 import type { LogLevel } from './types';
 
 // truecolor because a terminal theme remaps chalk's 16-color names (blue turns orange in monokai)
@@ -10,12 +8,3 @@ export const LEVEL_COLOR: Record<LogLevel, string> = {
     debug: '#66b3ff',
     trace: '#b399e6'
 };
-
-export const paint = {
-    sky: chalk.hex('#8fc7ff'),
-    iris: chalk.hex('#e29bff'),
-    mint: chalk.hex('#66d98a'),
-    amber: chalk.hex('#ffc061'),
-    coral: chalk.hex('#ff6b85'),
-    mute: chalk.dim
-} as const;
